divide into 2 scripts for code collaborative analysis and its helper
